requirements.txt
colorama>=0.4.6
flask>=3.1.0
python-chess>=1.999
pytest>=9.0.0
